chore: 配置调整
