+++ packages/plugin-auto-navbar/package.json
@@ -4,7 +4,14 @@
   "main": "src/node/index.ts",
   "author": "with-rainbow",
   "license": "MIT",
+  "scripts": {
+    "clean": "tsc --build --clean"
+  },
   "devDependencies": {
-    "@vuepress/core": "^1.9.7"
+    "@vuepress/core": "^1.9.7",
+    "@vuepress/theme-default": "^1.9.7",
+    "vite": "^2.9.8",
+    "vitest": "^0.10.2",
+    "vuepress": "2.0.0-beta.43"
   }
 }
