+++ packages/plugin-auto-navbar/tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "CommonJS",
+    "rootDir": "./src",
+    "outDir": "./lib"
+  },
+  "include": [
+    "./src"
+  ],
+  "references": []
+}
