+++ tsconfig.build.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.base.json",
   "references": [
-    "./packages/plugin-auto-navbar/tsconfig.build.json",
+    "./packages/plugin-auto-navbar",
   ],
   "files": []
 }
